[Fix] 모듈화 과정에서 정상적으로 작동하지 않는 signin페이지 정상작동하도록 수정
