+++ js/signin.js
@@ -1,19 +1,8 @@
 import { isVaildEmail } from "./utils/isValidEmail.js";
 import { querySelector } from "./utils/querySelector.js";
-import { updateErrorMessage } from "./functions/updateErrorMessage.js";
-import { passwordShowHidden } from "./functions/passwordShowHidden.js";
-import { checkValidLogin } from "./functions/CheckValidLogin.js";
-import {
-  passwordEyeIcon,
-  checkPasswordValidity,
-  tryLogin,
-} from "./utils/signCommons.js";
-import {
-  signinEmailInput,
-  signinPasswordInput,
-  loginBtn,
-  loginForm,
-} from "./utils/commonsBase.js";
+import { updateErrorMessage } from "./utils/updateErrorMessage.js";
+import { passwordShowHidden } from "./utils/passwordShowHidden.js";
+import { checkValidLogin } from "./utils/CheckValidLogin.js";
 
 /** 이메일 입력 값이 유효한 지 확인하는 함수 */
 function checkEmail(event) {
@@ -42,6 +31,38 @@ function checkEmail(event) {
   }
 }
 
+function tryLogin(checkValidLogin) {
+  const emailValue = signinEmailInput.value;
+  const passwordValue = signinPasswordInput.value;
+  const errorMessage = querySelector("#email-error-message");
+  const errorMessageBox = querySelector("#password-error-message");
+
+  const loginResult = checkValidLogin(emailValue, passwordValue);
+  if (loginResult === 1) {
+    location.href = "folder.html";
+  } else if (loginResult === -1) {
+    updatePasswordErrorMessage(
+      errorMessageBox,
+      "비밀번호를 확인해주세요",
+      signinPasswordInput,
+      true
+    );
+  } else {
+    updateEmailErrorMessage(
+      errorMessage,
+      "이메일을 확인해주세요.",
+      signinEmailInput,
+      true
+    );
+    updatePasswordErrorMessage(
+      errorMessageBox,
+      "비밀번호를 확인해주세요",
+      signinEmailInput,
+      true
+    );
+  }
+}
+
 /** 비밀번호 입력 유무 확인 */
 function checkPassword(event) {
   const passwordValue = event.target.value;
@@ -53,38 +74,6 @@ function checkPassword(event) {
     signinPasswordInput,
     passwordValue === ""
   );
-}
-
-/** 로그인 시도 */
-function tryLogin() {
-  const emailValue = signinEmailInput.value;
-  const passwordValue = signinPasswordInput.value;
-  const errorMessage = querySelector("#email-error-message");
-  const errorMessageBox = querySelector("#password-error-message");
-
-  if (checkValidLogin(emailValue, passwordValue) === 1) {
-    location.href = "folder.html";
-  } else if (checkValidLogin(emailValue, passwordValue) === -1) {
-    updateErrorMessage(
-      errorMessageBox,
-      "비밀번호를 확인해주세요",
-      signinPasswordInput,
-      true
-    );
-  } else {
-    updateErrorMessage(
-      errorMessage,
-      "이메일을 확인해주세요.",
-      signinEmailInput,
-      true
-    );
-    updateErrorMessage(
-      errorMessageBox,
-      "비밀번호를 확인해주세요",
-      signinPasswordInput,
-      true
-    );
-  }
 }
 
 // 이벤트 등록을 위한 변수 설정
