+++ js/signup.js
@@ -1,8 +1,8 @@
 import { isVaildEmail } from "./utils/isValidEmail.js";
 import { querySelector } from "./utils/querySelector.js";
-import { updateErrorMessage } from "./functions/updateErrorMessage.js";
-import { passwordShowHidden } from "./functions/passwordShowHidden.js";
-import { validEmail } from "./functions/CheckValidLogin.js";
+import { updateErrorMessage } from "./utils/updateErrorMessage.js";
+import { passwordShowHidden } from "./utils/passwordShowHidden.js";
+import { validEmail } from "./utils/CheckValidLogin.js";
 import { isVaildPassword } from "./utils/isValidPassword.js";
 
 /** 이메일 입력 값이 유효한 지 확인하는 함수 */
