+++ js/utils/checkValidSignup.js
